refactor out socket.io
